requirements.txt
scikit-learn
tqdm
pandas
gradio
pydantic==2.9.2
openai==1.61.0
langchain_core==0.3.22
langchain_openai==0.2.11
langchain_anthropic==0.2.3
numpy==1.26.4
scipy==1.13.1
langchain==0.3.7
langgraph==0.2.39
langchain_experimental==0.3.3
ipython
python-dotenv